req_mano_obra, req_materiales, req_equipos, req_fondos_avance, modified name routes, registros.php: modified name routes
