+++ app/Http/Controllers/Front_End/Constructor/RequerimientoRecursosController.php
@@ -3,22 +3,34 @@
 namespace App\Http\Controllers\Front_End\Constructor;
 
 use App\Http\Controllers\Controller;
+use App\Models\Constructor\Requerimiento;
+use App\Models\Constructor\RequerimientoItem;
+use App\Models\Constructor\RequerimientoRecursos;
 use Illuminate\Http\Request;
 
-class RequerimientoMaterialesController extends Controller
+class RequerimientoRecursosController extends Controller
 {
     public function inicio()
     {
+        return view('front-end.constructor.IndexManoObra');
+    }
+    public function inicioMaterial(){
         return view('front-end.constructor.IndexMateriales');
+    }
+    public function inicioEquipos(){
+        return view('front-end.constructor.IndexEquipos');
+    }
+    public function inicioFondosAvance(){
+        return view('front-end.constructor.IndexFondosAvance');
     }
     /**
      * Display a listing of the resource.
      *
-     * @return \Illuminate\Http\Response
+     * @return RequerimientoItem[]|\Illuminate\Database\Eloquent\Collection|\Illuminate\Http\Response
      */
     public function index()
     {
-        //
+        return RequerimientoRecursos::all();
     }
 
     /**
@@ -76,11 +88,32 @@ class RequerimientoMaterialesController extends Controller
      *
      * @param  \Illuminate\Http\Request  $request
      * @param  int  $id
-     * @return \Illuminate\Http\Response
+     * @return bool
      */
     public function update(Request $request, $id)
     {
-        //
+        $req_mano_obra = RequerimientoRecursos::findOrFail($id);
+        return $req_mano_obra->update([
+                'tipo_requerimiento_id' => $request->modal_tipo_requerimiento,
+                'codigo_recurso' => $request->modal_codigo,
+                'descripcion_recurso' => $request->modal_descripcion,
+                'unidad_id' => $request->modal_unidad,
+                'precio_referencial' => $request->modal_precio_referencial,
+                'unidad_contrato' => $request->modal_unidad_contrato,
+            ]
+        );
+    }
+    public function updateRequerimientoRecurso(Request $request, $id){
+        $req_mano_obra = RequerimientoRecursos::findOrFail($id);
+        return $req_mano_obra->update([
+                'tipo_requerimiento_id' => $request->modal_tipo_requerimiento,
+                'codigo_recurso' => $request->modal_codigo,
+                'descripcion_recurso' => $request->modal_descripcion,
+                'unidad_id' => $request->modal_unidad,
+                'precio_referencial' => $request->modal_precio_referencial,
+                'unidad_contrato' => $request->modal_unidad_contrato,
+            ]
+        );
     }
 
     /**
@@ -91,6 +124,8 @@ class RequerimientoMaterialesController extends Controller
      */
     public function destroy($id)
     {
-        //
+        $requerimientoRecurso = RequerimientoRecursos::findOrFail($id);
+        $requerimientoRecurso->delete();
+        return $requerimientoRecurso;
     }
 }
