+++ routes/otras_rutas/front-end/constructor/Registros.php
@@ -31,13 +31,12 @@ Route::group(['prefix' => 'paginas', 'namespace' => 'Front_End\Constructor'], fu
     Route::post('procesar_csv', 'PlanillaController@procesarCSV')->name('procesar_csv');
 
 //requerimientos
-    Route::apiResource('requerimiento_mano_obra', RequerimientoManoObraController::class);
-    Route::get('req_mano_obra', 'RequerimientoManoObraController@inicio')->name('req_mano_obra');
-    Route::get('req_materiales', 'RequerimientoManoObraController@inicioMaterial')->name('req_materiales');
-    Route::get(' req_equipos', 'RequerimientoManoObraController@inicioEquipos')->name('req_equipos');
-    Route::get('req_fondos_avance', 'RequerimientoManoObraController@inicioFondosAvance')->name('req_fondos_avance');
-    Route::post('update_req_mano_obra/{id}', 'RequerimientoManoObraController@updateManoObra')->name('update_req_mano_obra');
-    Route::apiResource('requerimiento_materiales', RequerimientoMaterialesController::class);
+    Route::apiResource('requerimiento_recurso', RequerimientoRecursosController::class);
+    Route::get('req_mano_obra', 'RequerimientoRecursosController@inicio')->name('req_mano_obra');
+    Route::get('req_materiales', 'RequerimientoRecursosController@inicioMaterial')->name('req_materiales');
+    Route::get(' req_equipos', 'RequerimientoRecursosController@inicioEquipos')->name('req_equipos');
+    Route::get('req_fondos_avance', 'RequerimientoRecursosController@inicioFondosAvance')->name('req_fondos_avance');
+    Route::post('update_req_recurso/{id}', 'RequerimientoRecursosController@updateRequerimientoRecurso')->name('update_req_mano_obra');
 //    Route::get('req_materiales', 'RequerimientoMaterialesController@inicio')->name('req_materiales');
 
     Route::apiResource('requerimientos', RequerimientoController::class);
